Separated routing module
Separated routing module from main root module
