+++ src/app/app.module.ts
@@ -1,4 +1,4 @@
-import { Component, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppComponent } from './app.component';
@@ -6,20 +6,7 @@ import { ExpenseCalculatorComponent } from './expense-calculator/expense-calcula
 import { RetirementCorpusCalculatorComponent } from './retirement-corpus-calculator/retirement-corpus-calculator.component';
 import { MonthlyInvestmentsCalculatorComponent } from './monthly-investments-calculator/monthly-investments-calculator.component';
 import { AppNavigationComponent } from './app-navigation/app-navigation.component';
-import { RouterModule, Routes } from '@angular/router';
-
-const routes: Routes = [
-  { path: 'expense-calc', component: ExpenseCalculatorComponent },
-  {
-    path: 'retirement-corpus-calc',
-    component: RetirementCorpusCalculatorComponent,
-  },
-  {
-    path: 'monthly-invst-calc',
-    component: MonthlyInvestmentsCalculatorComponent,
-  },
-  { path: '**', redirectTo: 'expense-calc' },
-];
+import { AppRoutingModule } from './app.routing.module';
 
 @NgModule({
   declarations: [
@@ -29,7 +16,7 @@ const routes: Routes = [
     MonthlyInvestmentsCalculatorComponent,
     AppNavigationComponent,
   ],
-  imports: [BrowserModule, RouterModule.forRoot(routes)],
+  imports: [BrowserModule, AppRoutingModule],
   providers: [],
   bootstrap: [AppComponent],
 })
