+++ src/app/expense-calculator/expense-calculator.component.html
@@ -1,38 +1,38 @@
 <div class="card shadow-sm p-3 mb-5 bg-body rounded">
-    <div class="card-header display-6">
-        Expenses at Retirement Calculator
-    </div>
-    <div class="card-body">
-        <form>
-            <div class="mb-3">
-                <label for="cur-mnth-exp" class="form-label">Current Monthyl Expense</label>
-                <input type="text" class="form-control" id="cur-mnth-exp">
-            </div>
-            <div class="mb-3">
-                <label for="cur-annl-exp" class="form-label">Current Annual Expense</label>
-                <input type="text" class="form-control" id="cur-annl-exp">
-            </div>
-            <div class="mb-3">
-                <label for="annl-inf-rte" class="form-label">Annual Inflation Rate</label>
-                <input type="text" class="form-control" id="annl-inf-rte">
-            </div>
-            <div class="mb-3">
-                <label for="curr-age" class="form-label">Current Age(yrs)</label>
-                <input type="text" class="form-control" id="curr-age">
-            </div>
-            <div class="mb-3">
-                <label for="exp-retir-age" class="form-label">Expected Retirement Age(yrs)</label>
-                <input type="text" class="form-control" id="exp-retir-age">
-            </div>
-            <div class="mb-3">
-                <label for="yrs-of-retire" class="form-label">Years of Retirement</label>
-                <input type="text" class="form-control" id="yrs-of-retire">
-            </div>
-            <div>
-                <label class="fw-bold text-primary">Expense at Retirement (Inflation Adjusted)</label>
-                <p class="fw-bold text-primary">100.2</p>
-            </div>
-            <button type="submit" class="btn btn-primary">Calculate Retirement Corpus... </button>
-        </form>
-    </div>
+  <div class="card-header display-6">
+    Expenses at Retirement Calculator
+  </div>
+  <div class="card-body">
+    <form>
+      <div class="form-floating mb-3">
+        <input type="text" class="form-control" id="cur-mnth-exp" placeholder="00.00">
+        <label for="cur-mnth-exp">Current Monthyl Expense</label>
+      </div>
+      <div class="form-floating mb-3">
+        <input type="text" class="form-control" id="cur-annl-exp" placeholder="00.00">
+        <label for="cur-annl-exp">Current Annual Expense</label>
+      </div>
+      <div class="form-floating mb-3">
+        <input type="text" class="form-control" id="annl-inf-rte" placeholder="0">
+        <label for="annl-inf-rte">Annual Inflation Rate</label>
+      </div>
+      <div class="form-floating mb-3">
+        <input type="text" class="form-control" id="curr-age" placeholder="0">
+        <label for="curr-age">Current Age(yrs)</label>
+      </div>
+      <div class="form-floating mb-3">
+        <input type="text" class="form-control" id="exp-retir-age" placeholder="0">
+        <label for="exp-retir-age">Expected Retirement Age(yrs)</label>
+      </div>
+      <div class="form-floating mb-3">
+        <input type="text" class="form-control" id="yrs-of-retire" placeholder="0">
+        <label for="yrs-of-retire">Years of Retirement</label>
+      </div>
+      <div>
+        <label class="fw-bold text-primary">Expense at Retirement (Inflation Adjusted)</label>
+        <p class="fw-bold text-primary">100.2</p>
+      </div>
+      <button type="submit" class="btn btn-primary">Calculate Retirement Corpus... </button>
+    </form>
+  </div>
 </div>
